refactor(html): drop the -media suffix from the Mux flavor modules
The React components are `hls-js` and `spf`; the html elements were `hls-js-media` and
`spf-media`, so documenting a flavor meant naming it two ways depending on the platform. Both
packages now spell it the same.

Internal module names only. The published `media/*` subpaths resolve to the define entries, so
nothing a consumer imports moves.
