+++ packages/html/src/media/mux-audio/index.ts
@@ -1,1 +1,1 @@
-export * from './hls-js-media';
+export * from './hls-js';
