+++ packages/html/src/media/mux-audio/spf.ts
@@ -8,7 +8,7 @@ const MuxAudioBase = MuxAudioMixin(MediaAttachMixin(CustomMediaElement('audio', 
 /**
  * `<mux-audio>` over the SPF audio-only Mux Media instead of the hls.js-backed one.
  *
- * Shares its name with the flavor in `./hls-js-media` on purpose: the import path picks
+ * Shares its name with the flavor in `./hls-js` on purpose: the import path picks
  * the engine, and nothing else about the surface moves. Deliberately not exported
  * from this directory's barrel, so importing one flavor never pulls the other's
  * engine in with it.
